+++ src/Type/AnimationType.php
@@ -74,4 +74,11 @@ class AnimationType
      * @var int|null
      */
     public $fileSize;
+
+    /**
+     * Optional. Thumbnail of the file, as defined by the sender.
+     *
+     * @var PhotoSizeType|null
+     */
+    public $thumbnail;
 }