+++ src/Type/AudioType.php
@@ -74,4 +74,11 @@ class AudioType
      * @var PhotoSizeType|null
      */
     public $thumb;
+
+    /**
+     * Optional. Thumbnail of the file, as defined by the sender.
+     *
+     * @var PhotoSizeType|null
+     */
+    public $thumbnail;
 }